+++ test/layers/msgpass.jl
@@ -2,7 +2,8 @@ in_channel = 10
 out_channel = 5
 num_V = 6
 num_E = 7
-adj = [0. 1. 0. 0. 0. 0.;
+T = Float32
+adj = T[0. 1. 0. 0. 0. 0.;
        1. 0. 0. 1. 1. 1.;
        0. 0. 0. 0. 0. 1.;
        0. 1. 0. 0. 1. 0.;
@@ -16,8 +17,9 @@ NewLayer(m, n) = NewLayer(randn(m,n))
 
 (l::NewLayer)(fg) = propagate(l, fg, :add)
 
-X = Array(reshape(1:num_V*in_channel, in_channel, num_V))
+X = Array{T}(reshape(1:num_V*in_channel, in_channel, num_V))
 fg = FeaturedGraph(adj, X)
+fg.ef = Fill(zero(T), 0, 2num_E)
 
 l = NewLayer(out_channel, in_channel)
 
@@ -25,7 +27,7 @@ l = NewLayer(out_channel, in_channel)
     @testset "no message or update" begin
         fg_ = l(fg)
 
-        @test graph(fg_) === adj
+        @test graph(fg_) == adj
         @test size(node_feature(fg_)) == (in_channel, num_V)
         @test size(edge_feature(fg_)) == (in_channel, 2*num_E)
         @test size(global_feature(fg_)) == (0,)
@@ -35,7 +37,7 @@ l = NewLayer(out_channel, in_channel)
     @testset "message function" begin
         fg_ = l(fg)
 
-        @test graph(fg_) === adj
+        @test graph(fg_) == adj
         @test size(node_feature(fg_)) == (out_channel, num_V)
         @test size(edge_feature(fg_)) == (out_channel, 2*num_E)
         @test size(global_feature(fg_)) == (0,)
@@ -45,7 +47,7 @@ l = NewLayer(out_channel, in_channel)
     @testset "message and update" begin
         fg_ = l(fg)
 
-        @test graph(fg_) === adj
+        @test graph(fg_) == adj
         @test size(node_feature(fg_)) == (out_channel, num_V)
         @test size(edge_feature(fg_)) == (out_channel, 2*num_E)
         @test size(global_feature(fg_)) == (0,)