+++ test/runtests.jl
@@ -2,6 +2,7 @@ using GeometricFlux
 using GeometricFlux.Datasets
 using Flux
 using Flux: @functor
+using FillArrays
 using GraphSignals
 using StaticArrays: @MMatrix, @MArray
 using LightGraphs: SimpleGraph, SimpleDiGraph, add_edge!, nv, ne
@@ -19,7 +20,6 @@ cuda_tests = [
 ]
 
 tests = [
-    "datasets/datasets",
     "layers/gn",
     "layers/msgpass",
     "layers/conv",